Move ai costs around
